Trim and refactor
